+++ HSTracker/Core/SizeHelper.swift
@@ -380,33 +380,6 @@ struct SizeHelper {
         return result
     }
     
-    static func mercenariesButtonOffset() -> Double {
-        let h = hearthstoneWindow.height
-        if AppDelegate.instance().coreManager.game.isInMenu && screenRatio > 0.9 {
-            return h * 0.104
-        }
-        return h * 0.05
-    }
-    
-    static func mercenariesTaskListButton() -> NSRect {
-        let w = 150.0
-        let h = 60.0
-        let height = hearthstoneWindow.height
-        let bottom = hearthstoneWindow.frame.minY + mercenariesButtonOffset()
-        let right = hearthstoneWindow.frame.maxX - height * 0.01
-        let frame = NSRect(x: right - w, y: bottom, width: w, height: h)
-        return frame
-    }
-    
-    static func mercenariesTaskListView() -> NSRect {
-        let frame = mercenariesTaskListButton()
-        let height = hearthstoneWindow.height
-        let width = hearthstoneWindow.width / 2.0
-        let bottom = hearthstoneWindow.frame.minY + frame.height + mercenariesButtonOffset() + 8
-        
-        return NSRect(x: frame.maxX - width, y: bottom, width: width, height: height - bottom)
-    }
-    
     static func flavorTextFrame() -> NSRect {
         let hs = hearthstoneWindow.frame