+++ HSTracker/UIs/Trackers/WindowManager.swift
@@ -57,14 +57,6 @@ class WindowManager {
         return $0
     }(BoardOverlay(windowNibName: "BoardOverlay"))
     
-    var mercenariesTaskListButton: MercenariesTaskListButton = {
-        return $0
-    }(MercenariesTaskListButton(windowNibName: "MercenariesTaskListButton"))
-
-    var mercenariesTaskListView: MercenariesTaskListView = {
-        return $0
-    }(MercenariesTaskListView(windowNibName: "MercenariesTaskListView"))
-    
     var flavorText: FlavorText = {
         return $0
     }(FlavorText(windowNibName: "FlavorText"))